fix code using moved tx ctx
