+++ crates/tests/Cargo.toml
@@ -30,7 +30,7 @@ namada-eth-bridge = [
 namada_core = {path = "../core", features = ["testing"]}
 namada_sdk = {path = "../sdk", default-features=false, features = ["download-params", "testing", "wasm-runtime"]}
 namada_test_utils = {path = "../test_utils"}
-namada_tx_env = {path = "../tx_env"}
+namada_tx_env = {path = "../tx_env", features = ["testing"]}
 namada_tx_prelude = {path = "../tx_prelude"}
 namada_vp = {path = "../vp"}
 namada_vp_prelude = {path = "../vp_prelude"}
