+++ crates/tests/src/lib.rs
@@ -6,7 +6,7 @@
 #![deny(rustdoc::private_intra_doc_links)]
 
 mod vm_host_env;
-pub use vm_host_env::{tx, vp};
+pub use vm_host_env::{tx_env, vp};
 #[cfg(test)]
 mod e2e;
 #[cfg(test)]
