+++ crates/vm/src/wasm/mod.rs
@@ -8,3 +8,4 @@ pub mod run;
 pub use compilation_cache::common::{Cache, CacheName};
 pub use compilation_cache::tx::TxCache;
 pub use compilation_cache::vp::VpCache;
+pub use wasmer;
