+++ wasm/tx_claim_rewards/src/lib.rs
@@ -6,7 +6,7 @@ use namada_tx_prelude::*;
 #[transaction]
 fn apply_tx(ctx: &mut Ctx, tx_data: BatchedTx) -> TxResult {
     let data = ctx.get_tx_data(&tx_data)?;
-    let withdraw = transaction::pos::Withdraw::try_from_slice(&data[..])
+    let withdraw = data::pos::Withdraw::try_from_slice(&data[..])
         .wrap_err("Failed to decode Withdraw value")?;
 
     ctx.claim_reward_tokens(withdraw.source.as_ref(), &withdraw.validator)
